Changes to time measurements: - calculate time in more type-safe way - don't store scope depth - store frame durations in microseconds

+++ src/Utils/Profiling/ProfileDataWriter.h
@@ -17,7 +17,7 @@ public:
 		ScopedProfilerThreadData::Records records;
 	};
 	using ScopedProfilerDatas = std::vector<ScopedProfilerData>;
-	using FrameDurations = std::vector<long>;
+	using FrameDurations = std::vector<double>;
 
 	struct ProfileData
 	{
